microservices/user_service/requirements.txt
# Django and core dependencies
Django>=4.2.0,<5.0.0
django-environ>=0.10.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0

# Celery and task queue
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Database and caching
redis>=4.5.0
django-redis>=5.2.0

# API documentation
drf-spectacular>=0.26.0

# Validation and utilities
Pillow>=9.5.0
requests>=2.28.0
python-dateutil>=2.8.0

# Monitoring and logging
django-prometheus>=2.2.0
sentry-sdk>=1.25.0
django-health-check>=3.17.0
django-extensions>=3.2.0

# Development and testing
pytest>=7.3.0
pytest-django>=4.5.0
factory-boy>=3.2.0
faker>=18.0.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Security
django-ratelimit>=4.0.0
django-guardian>=2.4.0